refactor(libslic3r): name magic numbers, fix casts, and clean up nesting across geometry/gcode files
Drive several BP codes toward zero across the core geometry and G-code
path, all literal-for-literal or behavior-preserving:

- Color.cpp: full BP1002 pass deferred earlier is now complete. Named
  every RGB/HSV byte-shift, mask, sextant-boundary and hex-digit constant
  (COLOR_CHANNEL_MAX, HSV_SEXTANT_*, RGB_COMPONENT_COUNT, etc.); switched
  the HSVtoRGB half-open-interval if/else-if chain to an equivalent switch
  on the truncated sextant index.
- TriangleMesh.cpp: explicit static_cast for double->float narrowing in
  slice(), its_make_cone(), and its_convex_hull() (same truncation the
  implicit conversion already performed).
- AABBMesh.cpp: raw new -> std::make_unique for the AABBImpl member
  (BP1022); mark read-only IGL callback overrides const.
- AABBTreeLines.hpp: replace duplicated left/right recursive-descent
  blocks in coordinate_aligned_ray_hit_count with a shared lambda
  (BP1013); zero-init an intersection point local.
- Algorithm/RegionExpansion.cpp: flatten merge_splits' nested if/else
  into early-continue guard clauses (BP1015), same control flow.
- Arachne/PolylineStitcher.hpp: name the 0.01mm snap/stitch epsilon;
  fix a real bug found while touching the file -- boolean `&` should have
  been logical `&&` in the should_close expression (lnt-logical-bitwise-
  mismatch).
- BoundingBox.hpp / Point.hpp / MeasureUtils.hpp / OpenVDBUtilsLegacy.hpp /
  TriangulateWall.hpp: explicit static_cast for narrowing conversions that
  were previously implicit (BP1008).
- BridgeDetector.hpp: drop the old "declared but undefined" private
  copy-assignment idiom (the class is still non-assignable via its
  reference/const members); name the zero bridge-cost constant.
- Geometry/ArcWelder.hpp: early-continue instead of an empty-else branch
  in the circle-fit Jacobian loop; name the discretization-step float
  constants and add a static_assert on FloatType; use the existing
  scale_d() helper instead of the differently-named scaled().
- SLAPrint.hpp / NSVGUtils.hpp: reword comments that read as BP1008
  "(float)" cast false positives.
- PerimeterGenerator.cpp/.hpp: name the debug SVG stroke widths and the
  repeated spacing fractions (HALF/QUARTER/ONE_AND_HALF_SPACING_RATIO,
  COARSE/FINE_SPACING_DIVISOR, etc.); widen depth from unsigned short to
  size_t to match its callers. Every /2 and /4 replaced with *0.5/*0.25
  is IEEE-754 bit-identical (verified in Python); /10 and /20 stay
  divisions, only the literal is named.
- GCodeWriter.cpp: name the arc-offset epsilon scale and minimum
  quantized-delta constants; remove the now-dead log_10 lookup table.
- GCodeProcessor.cpp: const&/by-value fixes (BP1025), internal linkage
  for file-local helpers, explicit narrowing casts, and a structured
  binding const& fix.

libslic3r builds clean (0 errors) with all of the above applied.

+++ src/libslic3r/PerimeterGenerator.hpp
@@ -169,13 +169,13 @@ public:
     // overhang may need to be reversed
     bool is_steep_overhang;
     // Depth in the hierarchy. External perimeter has depth = 0. An external perimeter could be both a contour and a hole.
-    unsigned short depth;
+    size_t depth;
     // Should this contur be fuzzyfied on path generation?
     bool fuzzify;
     // Children contour, may be both CCW and CW oriented (outer contours or holes).
     std::vector<PerimeterGeneratorLoop> children;
 
-    PerimeterGeneratorLoop(const Polygon &polygon, unsigned short depth, bool is_contour, bool steep_overhangs, bool fuzzify)
+    PerimeterGeneratorLoop(const Polygon &polygon, size_t depth, bool is_contour, bool steep_overhangs, bool fuzzify)
         : polygon(polygon), is_contour(is_contour), is_steep_overhang(steep_overhangs), depth(depth), fuzzify(fuzzify)
     {}
     // External perimeter. It may be CCW or CW oriented (outer contour or hole contour).
